+++ applications/MappingApplication/tests/mapper_test_case.py
@@ -4,6 +4,9 @@ import KratosMultiphysics as KM
 from KratosMultiphysics import KratosUnittest
 data_comm = KM.DataCommunicator.GetDefault()
 import os
+from KratosMultiphysics import from_json_check_result_process
+from KratosMultiphysics import json_output_process
+from KratosMultiphysics import vtk_output_process
 
 def GetFilePath(fileName):
     return os.path.join(os.path.dirname(os.path.realpath(__file__)), "mdpa_files", fileName)
@@ -29,9 +32,12 @@ class MapperTestCase(KratosUnittest.TestCase):
         # list of variables involved in the Mapper-Tests
         cls.model_part_origin.AddNodalSolutionStepVariable(KM.PRESSURE)
         cls.model_part_origin.AddNodalSolutionStepVariable(KM.FORCE)
+        cls.model_part_origin.AddNodalSolutionStepVariable(KM.DISPLACEMENT)
 
         cls.model_part_destination.AddNodalSolutionStepVariable(KM.TEMPERATURE)
         cls.model_part_destination.AddNodalSolutionStepVariable(KM.VELOCITY)
+        cls.model_part_destination.AddNodalSolutionStepVariable(KM.REACTION)
+        cls.model_part_destination.AddNodalSolutionStepVariable(KM.MESH_DISPLACEMENT)
 
         cls.model_part_origin.ProcessInfo[KM.DOMAIN_SIZE] = 3 # needed for the partitioner!
         cls.model_part_destination.ProcessInfo[KM.DOMAIN_SIZE] = 3 # needed for the partitioner!
@@ -104,3 +110,61 @@ def ReadDistributedModelPart(model_part, mdpa_file_name):
     model_part_import_util = trilinos_import_model_part_utility.TrilinosImportModelPartUtility(model_part, importer_settings)
     model_part_import_util.ImportModelPart()
     model_part_import_util.CreateCommunicators()
+
+def OutputReferenceSolution(model_part, variable, file_name):
+    full_model_part_name = model_part.Name
+    if model_part.IsSubModelPart():
+        full_model_part_name = model_part.GetParentModelPart().Name  + "." + full_model_part_name
+
+    if data_comm.IsDistributed():
+        raise Exception("Writing of reference results in not possible in MPI!")
+    KM.Logger.PrintWarning('BladeMappingTests', 'Writing reference solution for ModelPart "{}"; Variable "{}"; FileName "{}"'.format(full_model_part_name, variable.Name(), file_name))
+
+    output_parameters = KM.Parameters("""{
+        "output_variables"     : [\"""" + variable.Name() + """\"],
+        "output_file_name"     : \"""" + file_name + """\",
+        "model_part_name"      : \"""" + full_model_part_name + """\",
+        "time_frequency"       : 0.00,
+        "use_node_coordinates" : true
+    }""")
+
+    output_proc = json_output_process.JsonOutputProcess(model_part.GetModel(), output_parameters)
+    output_proc.ExecuteInitialize()
+    output_proc.ExecuteBeforeSolutionLoop()
+    output_proc.ExecuteFinalizeSolutionStep()
+
+def CheckHistoricalNonUniformValues(model_part, variable, file_name, output_reference_solution=False):
+    if output_reference_solution:
+        OutputReferenceSolution(model_part, variable, file_name)
+    else:
+        full_model_part_name = model_part.Name
+        if model_part.IsSubModelPart():
+            full_model_part_name = model_part.GetParentModelPart().Name  + "." + full_model_part_name
+
+        check_parameters = KM.Parameters("""{
+            "check_variables"           : [\"""" + variable.Name() + """\"],
+            "input_file_name"           : \"""" + file_name + """\",
+            "model_part_name"           : \"""" + full_model_part_name + """\",
+            "tolerance"                 : 1e-6,
+            "relative_tolerance"        : 1e-9,
+            "time_frequency"            : 0.00,
+            "use_node_coordinates"      : true,
+            "check_only_local_entities" : true
+        }""")
+        # TODO check all entities, requires some syncronization though!
+        check_proc = from_json_check_result_process.FromJsonCheckResultProcess(model_part.GetModel(), check_parameters)
+        check_proc.ExecuteInitialize()
+        check_proc.ExecuteFinalizeSolutionStep()
+
+def VtkOutputNodesHistorical(model_part, variable, prefix=""):
+    vtk_parameters = KM.Parameters("""{
+        "model_part_name"                    : \"""" + model_part.Name + """\",
+        "file_format"                        : "binary",
+        "output_control_type"                : "step",
+        "output_sub_model_parts"             : false,
+        "save_output_files_in_folder"        : false,
+            "custom_name_prefix"             : \"""" + prefix + "_" + """\",
+        "nodal_solution_step_data_variables" : [\"""" + variable.Name() + """\"]
+    }""")
+
+    vtk_output_process.VtkOutputProcess(model_part.GetModel(),vtk_parameters).PrintOutput()
